+++ src/main/java/com/fomdev/awaken/network/DifficultyHandler.java
@@ -4,6 +4,8 @@ import com.fomdev.awaken.difficulty.ClientDifficultyManager;
 import com.fomdev.awaken.packet.DifficultySyncPacketPayloadResponder;
 import net.neoforged.neoforge.network.handling.IPayloadContext;
 
+import java.math.BigDecimal;
+
 public class DifficultyHandler
 {
     public static void handleResponse(
@@ -12,7 +14,7 @@ public class DifficultyHandler
     )
     {
         context.enqueueWork(() -> {
-            float data = payload.difficulty();
+            BigDecimal data = payload.difficulty();
             ClientDifficultyManager.setDifficulty(data);
         });
     }
